requirements.txt
llama_cpp_python==0.2.64